reorg files and do first critical path of charge transaction
